requirements/production.txt
# Production requirements
-r base.txt

# Error monitoring and tracking
sentry-sdk>=1.35.0

# Structured logging (JSON format)
python-json-logger>=2.0.7

# File storage (S3)
django-storages>=1.14.0
boto3>=1.29.0

# Performance monitoring (opcional)
# newrelic>=9.2.0

# CORS (si se necesita API con frontend separado)
django-cors-headers>=4.3.0
